Reactivate Landing page and tests, lock more packages
