+++ .gitignore
@@ -1,1 +1,2 @@
 /acceptance_tests/junitxml/
+/node_modules
