+++ .prettierignore
@@ -1,0 +1,2 @@
+acceptance_tests/Pipfile.lock
+Pipfile.lock
