+++ Pipfile
@@ -7,9 +7,20 @@ name = "pypi"
 psycopg2 = "==2.9.1"
 numpy = "==1.21.3"
 pyyaml = "==6.0"
-pythonqwt = "==0.9.2"
+PythonQwt = "==0.9.2"
+owslib = "==0.25.0"
+jinja2 = "==3.0.3"
+pygments = "==2.11.2"
+pillow = "==9.0.0"
+shapely = "==1.8.0"
+future = "==0.18.2"
+termcolor = "==1.1.0"
 
 [dev-packages]
+geolinks = "==0.2.0 "
+six = "==1.16.0"
+nose2 = "==0.10.0"
+mock = "==4.3.0"
 
 [requires]
 python_version = "3.8"
